requirements.txt
openai
pandas
pyyaml
google-api-python-client
gspread
oauth2client
openpyxl
xlrd
